requirements.txt
torch
torchvision
opencv-python
numpy
PyYAML
Pillow
tqdm
